+++ integration-tests/tests/cirrus/mod.rs
@@ -8,7 +8,7 @@ fn should_run_restic_with_given_arguments() {
         .with_file("cirrus.toml", "");
     Command::cargo_bin("test-cirrus")
         .unwrap()
-        .arg("--restic-binary")
+        .arg("--restic")
         .arg(workdir.test_binary())
         .arg("--config-file")
         .arg(workdir.path().join("cirrus.toml"))
@@ -39,7 +39,7 @@ fn should_run_backup() {
     );
     Command::cargo_bin("test-cirrus")
         .unwrap()
-        .arg("--restic-binary")
+        .arg("--restic")
         .arg(workdir.test_binary())
         .arg("--config-file")
         .arg(workdir.path().join("cirrus.toml"))
@@ -65,7 +65,7 @@ fn should_run_restic_subcommand_without_config_file_if_possible() {
     let workdir = new_workdir();
     Command::cargo_bin("test-cirrus")
         .unwrap()
-        .arg("--restic-binary")
+        .arg("--restic")
         .arg(workdir.test_binary())
         .arg("--config-file")
         .arg(workdir.path().join("does-not-exist.toml"))