+++ CHANGELOG.md
@@ -4,9 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). Note that
 the version numbers are *not* semantic.
 
-## 1.3.0 - 2021-09-08
+## 1.3.0
 ### Changed
 * New tag format: `cirrus.<backup name>`
+* Changed `--restic-binary` argument to `--restic`.
 
 ## 1.2.1 - 2021-09-08
 ### Changed